Better pie legends
