+++ frontend/src/visualizations/components/Legend.css
@@ -3,6 +3,8 @@
 }
 
 :local .LegendItem {
+  font-size: 14px;
+  font-weight: bold;
   transition: opacity 0.25s linear;
   opacity: 1;
 }
@@ -17,10 +19,12 @@
 :local .Legend.vertical {
   display: flex;
   flex-direction: column;
+  margin-right: 1em;
 }
 
 :local .Legend.horizontal {
   display: flex;
   flex-direction: row;
   flex-wrap: wrap;
+  margin-top: 1em;
 }
